+++ internal/webhook/external/path.go
@@ -9,5 +9,5 @@ func GenerateWebhookPath(project, provider, token string) string {
 	input := []byte(project + provider + token)
 	h := sha256.New()
 	h.Write(input)
-	return fmt.Sprintf("/%x", h.Sum(nil))
+	return fmt.Sprintf("/webhook/%s/%x", provider, h.Sum(nil))
 }
